requirements.txt
PyYAML
gunicorn
Markdown
PyMySQL
Pillow
Django==2.2
djangorestframework==3.9.2